refactor(api): move scorecard model
