+++ api/tests/unit/domain/services/scorecard-service_test.js
@@ -2,9 +2,9 @@ import { expect, sinon, domainBuilder } from '../../../test-helper.js';
 import {
   Assessment,
   CompetenceEvaluation,
-  Scorecard,
   CampaignParticipationStatuses,
 } from '../../../../lib/domain/models/index.js';
+import { Scorecard } from '../../../../src/evaluation/domain/models/Scorecard.js';
 import * as scorecardService from '../../../../lib/domain/services/scorecard-service.js';
 import { KnowledgeElement } from '../../../../lib/domain/models/KnowledgeElement.js';
