+++ api/tests/unit/domain/usecases/find-tutorials_test.js
@@ -1,7 +1,7 @@
 import { sinon, expect, domainBuilder } from '../../../test-helper.js';
 import { UserNotAuthorizedToAccessEntityError } from '../../../../lib/domain/errors.js';
 import { KnowledgeElement } from '../../../../lib/domain/models/KnowledgeElement.js';
-import { Scorecard } from '../../../../lib/domain/models/Scorecard.js';
+import { Scorecard } from '../../../../src/evaluation/domain/models/Scorecard.js';
 import { findTutorials } from '../../../../lib/domain/usecases/find-tutorials.js';
 
 describe('Unit | UseCase | find-tutorials', function () {
