+++ api/tests/unit/domain/usecases/get-scorecard_test.js
@@ -1,6 +1,6 @@
 import { sinon, expect } from '../../../test-helper.js';
 import { UserNotAuthorizedToAccessEntityError } from '../../../../lib/domain/errors.js';
-import { Scorecard } from '../../../../lib/domain/models/Scorecard.js';
+import { Scorecard } from '../../../../src/evaluation/domain/models/Scorecard.js';
 import { getScorecard } from '../../../../lib/domain/usecases/get-scorecard.js';
 
 describe('Unit | UseCase | get-scorecard', function () {
