+++ api/tests/unit/domain/usecases/get-user-profile-shared-for-campaign_test.js
@@ -1,6 +1,6 @@
 import { expect, sinon, domainBuilder, catchErr } from '../../../test-helper.js';
 import { getUserProfileSharedForCampaign } from '../../../../lib/domain/usecases/get-user-profile-shared-for-campaign.js';
-import { Scorecard } from '../../../../lib/domain/models/Scorecard.js';
+import { Scorecard } from '../../../../src/evaluation/domain/models/Scorecard.js';
 import { NoCampaignParticipationForUserAndCampaign } from '../../../../lib/domain/errors.js';
 import { constants } from '../../../../lib/domain/constants.js';
